Update Trace doc
The `rgc` crate nominated in the old documentation was the name of a personal project containing both `rust-cc` (previously called rcc, which stands for Rust cycle Collector) and a mark-and-sweep gc (rgc stands for Rust Garbage Collector). Since the rgc name is already taken on crates.io and the mark-and-sweep gc wasn't as reliable as `rust-cc `, I decided to only publish the latter.
The old documentation was still referring to it, so I'm leaving a note here for curious people :wink:.