Incomplete attempt
Incomplete implementation of reworking fat pointers.
This is left intentionally incomplete, because it's been found out it is better to save the vtable only when calling `Cc::new` than to save it when tracing/dropping `Cc`s to increase the predictability of the library's behavior.

The use case for this feature is better supporting wrappers like `MaybeUninit`, which has an empty implementation of `trace` and `finalize`. When a `Cc<MaybeUninit<T>>` is casted to `Cc<T>`, the saved vtabled still doesn't trace anything, since it considers a `Cc<MaybeUninit<T>>`.
However, changing the approach to be completely dynamic is worse: given two `Cc` (pointing to the same allocation), one being a `Cc<MaybeUninit<T>>` and the other a `Cc<T>`, the actually-called trace/finalize/drop methods are almost unpredictable, since the saved vtable is the one of the first dropped or traced `Cc` between the two (and possibly many others, in a complex application).